test(frontend/copilot): split streaming tests into themed files + add coverage
Extract shared MSW + render helpers into sse-helpers.tsx so future tests
can reuse renderHost / typeAndSend / clickStop and the
copilotStreamSequenceHandler factory.

Split the single integration file by theme:
- copilotStream.content.test.tsx — text/reasoning/tool/multi-step
  rendering. Adds: two-turn continuity in the same session, and
  inline markdown emphasis (`**bold**` must not survive as literal).
- copilotStream.lifecycle.test.tsx — submit/stop button swap, stop
  mid-stream, resume on mount.
- copilotStream.errors.test.tsx — SSE error chunk, 429 rate-limit
  dialog, HTTP 500. Adds: partial text preserved when an error chunk
  arrives mid-stream, and the input stays enabled after HTTP 500 so
  the user can retry.

Net: 12 → 16 tests across 3 files (was 1).
